feat: check whether the user have the authority to modify the post

+++ src/main/java/xyz/aiinirii/postalk/service/PostService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import xyz.aiinirii.postalk.bean.Post;
+import xyz.aiinirii.postalk.bean.User;
 import xyz.aiinirii.postalk.mapper.PostMapper;
 import xyz.aiinirii.postalk.mapper.TextMapper;
 
@@ -56,23 +57,34 @@ public class PostService {
      * update the post
      *
      * @param post the post
+     * @param user the user
      * @return true if success
      */
     @Transactional(propagation = Propagation.REQUIRED)
-    public boolean updatePost(Post post) {
-        return postMapper.updatePost(post) == 1;
+    public boolean updatePost(Post post, User user) {
+        // check whether the user is the writer of the post
+        post.setUser(postMapper.findPostById(post.getId()).getUser());
+        if (post.getUser() == user) {
+            return postMapper.updatePost(post) == 1;
+        }
+        return false;
     }
 
     /**
      * delete the post by id
      *
-     * @param id the id
+     * @param id   the id
+     * @param user the user
      * @return true if success
      */
     @Transactional(propagation = Propagation.REQUIRED)
-    public boolean deletePostById(Integer id) {
-        int postDelete = postMapper.deletePostById(id);
-        int textDelete = textMapper.deleteTextById(id);
-        return postDelete == 1 && textDelete == 1;
+    public boolean deletePostById(Integer id, User user) {
+        // check whether the user is the writer of the post
+        if (postMapper.findPostById(id).getUser() == user) {
+            int postDelete = postMapper.deletePostById(id);
+            int textDelete = textMapper.deleteTextById(id);
+            return postDelete == 1 && textDelete == 1;
+        }
+        return false;
     }
 }
